refactor(music-folder): move music folder related endpoint to another folder
