+++ src/open_subsonic/mod.rs
@@ -5,6 +5,7 @@ pub mod extension;
 pub mod media_annotation;
 pub mod media_list;
 pub mod media_retrieval;
+pub mod music_folder;
 pub mod permission;
 pub mod scan;
 pub mod searching;
