+++ types/src/browsing/get_music_folders.rs
@@ -2,7 +2,7 @@ use nghe_proc_macros::{
     add_common_convert, add_request_types_test, add_subsonic_response, add_types_derive,
 };
 
-use super::MusicFolder;
+use crate::music_folder::MusicFolder;
 
 #[add_common_convert]
 pub struct GetMusicFoldersParams {}
