+++ types/src/browsing/mod.rs
@@ -3,13 +3,9 @@ pub mod get_album_info2;
 pub mod get_artist;
 pub mod get_artist_info2;
 pub mod get_artists;
-pub mod get_folder_stats;
 pub mod get_genres;
 pub mod get_indexes;
 pub mod get_music_directory;
 pub mod get_music_folders;
 pub mod get_song;
 pub mod get_top_songs;
-pub mod music_folder;
-
-pub use music_folder::{MusicFolder, MusicFolderPath};
